only set fetch's credentials:include option if no access token
this is what client.js does, and fixes a CORS issue when running client4.js in a browser.
Specifically, the issue seen is:

Fetch API cannot load http://127.0.0.1:8065/api/v4/users/me. Response to preflight request doesn't pass access control
check: The value of the 'Access-Control-Allow-Credentials' header in the response is '' which must be 'true' when the
request's credentials mode is 'include'. Origin 'http://127.0.0.1:8002' is therefore not allowed access.