MM-11556: introduce a getDirectAndGroupChannels
This is a simplified variant of `getDirectChannels` that does does not exclude manually or automatically closed direct and group channels, nor archived channels. The downstream webapp needs the whole set to display the channels in the "Other" section of the channel quick switcher.

Modifying `getDirectChannels` directly was out of scope:
* it is being used by the mobile app, and thus risks changing semantics unexpectedly
* it is extremely difficult to unravel the business logic. This is a candidate for significant refactoring.