MM-12233: pre-populate users in channel from posts
The idea here is to give clients of the Redux store an opportunity to know about users in the channel based solely on the post data that's been loaded so far. As noted in the comments, this won't correctly handle users that have been removed from the channel (and it would be non-trivial to figure out otherwise), but even those users might still make sense as "local" recommendations, despite their having left.