For terms of service, don't equate unknown with rejected.
Explicitly model a `loaded` parameter to distinguish the unknown state
from the loaded but rejected state, avoiding a race condition in
deciding that the terms of service should be shown when in fact we had
just loaded the user prior to fetching the accepted terms of service.

Ideally, the server would return the terms of service inline with the
user request (JOINing to the table in question, like we do for fetching
bot metadata).